+++ lib/src/view/uxutils/src/view/dialog_box.dart
@@ -73,7 +73,7 @@ Future<bool> showBox({
   }
   final ThemeData theme = Theme.of(context);
   final TextStyle dialogTextStyle =
-      theme.textTheme.subhead.copyWith(color: theme.textTheme.caption.color);
+      theme.textTheme.subtitle1.copyWith(color: theme.textTheme.caption.color);
   if (App.useMaterial) {
     return showDialog<bool>(
       context: context,